Refator: 사진 수정 부분 입력검증 모듈화
