+++ services/community/article.crud.community.service.js
@@ -11,6 +11,7 @@ import path from "path";
 import {
   addBaseUrl,
   deleteLocalFile,
+  isEditInputCorrect,
 } from "../../utils/upload/uploader.object.js";
 
 /**
@@ -201,79 +202,12 @@ export const updateArticle = async ({
     });
 
     // 이미지 순서 변경
-
-    /*
-        ** 이거 아님. 그냥 고민했던 흔적을 남기고 싶어서 남겨둠 **
-
-        imageSequence field를 받아서 수정을 진행합니다.
-        해당 field는 숫자들의 배열이며, 
-        각 숫자는 이미지 순서 및 삭제/추가 여부를 나타냅니다.
-
-        -1: 삭제된 이미지 (DB에서 삭제)
-        삭제된 이미지의 경우 기존 이미지 개수와 
-        동일한 길이의 배열이 들어왔을 때만 존재해야 합니다.
-        (지금 이거 예외 처리 안되있음, 그렇지 않으면 말이 안됨)
-        (한 번 종이에 써보면서 생각해보면 답 나옴)
-
-        -2: 새로 추가된 이미지 (DB에 추가)
-        imagePaths로 들어온 이미지 순서대로 넣어주면 됩니다.
-        첫 번째 -2 순서에 첫 번째로 들어온 이미지, ... 와 같이
-
-        나머지 숫자는 기존 이미지 순서를 나타냅니다.
-
-        ex) 원래 DB에 3개의 이미지가 있던 게시글에서
-        [2, 1, -2] 라는 배열이 들어온 경우
-        1. [기존 2번] 이미지를 1번으로 변경
-        2. [기존 1번] 이미지를 2번으로 변경
-        3. [새로 추가된 이미지]를 3번으로 추가
-
-        ex2) 원래 DB에 3개의 이미지가 있던 게시글에서
-        [2, -1, -1] 라는 배열이 들어온 경우
-        1. [기존 2번] 이미지를 1번으로 변경
-        2. [기존 1번, 3번] 이미지를 DB에서 삭제
-      */
-
-    /*
-        ** 이거 보면 됩니다 **
-
-        기존 이미지 Sequence와
-        Input으로 추가한 이미지의 Sequence를 
-        모두 받아서 하는 방향으로 수정하자.
-
-        existingImageSequence : 기존 파일들의 순서 및 삭제 여부
-        newImageSequence : 새로 추가된 파일들의 순서 (들어갈 곳)
-
-
-        동일하게 -1 삭제, 그 이외에는 수정하면 될듯.
-
-        existingImageSequence: [2, 1, -2]
-        newImageSequence: [4, 5, 6]
-
-        TODO : DB에 Unique key를 안 걸어둬서, 
-        eventId & sequence가 동일한 컬럼이 존재할 수 있지만 (가능하지만),
-        Unique key를 걸어서 중복을 방지함과 동시에 데이터 무결성을 지켜야 함.
-        아직 코드에선 해당 부분에 대한 예외 처리를 하지 않음.
-
-        (client가 잘못된 데이터를 보내는 경우에 대한 처리를 하지 않았다는 말)
-        근데 그냥 바로 해버림 6시 40분인데 자고 싶다
-      */
-
-    // 사용자가 보낸 이미지 순서가 이상한지 확인
-    const filteredExisting = existingImageSequence.filter((seq) => seq !== -1);
-    const combinedSequences = [...filteredExisting, ...newImageSequence];
-    const uniqueSequences = new Set(combinedSequences);
-
-    if (
-      existingImageSequence.length !== existingImages.length || // 존재하는 이미지 개수가 다른 경우
-      newImageSequence.length !== imagePaths.length || // 새로 추가된 이미지 개수가 다른 경우
-      uniqueSequences.size !== combinedSequences.length || // 중복된 순서가 있는 경우
-      Math.min(...combinedSequences) !== 1 || // 순서가 1부터 시작하지 않는 경우
-      Math.max(...combinedSequences) !== combinedSequences.length // 순서가 중간에 빠진 경우
-    ) {
-      throw new InvalidInputError(
-        "설명은 못하곘는데 이상한 입력 넣지 말아라 진짜"
-      );
-    }
+    isEditInputCorrect({
+      existingImageSequence,
+      newImageSequence,
+      existingImagesLength: existingImages.length,
+      newImageLength: imagePaths.length,
+    });
 
     // 기존 이미지 순서 변경
     existingImageSequence.map(async (seq, idx) => {
